+++ lib/src/platform_generic.dart
@@ -40,29 +40,31 @@ class DataSerializerPlatformGeneric extends DataSerializerPlatform {
   }
 
   @override
-  void setUint64(ByteData data, int n, [int offset = 0]) =>
-      setInt64(data, n, offset);
+  void setUint64(ByteData data, int n,
+          [int offset = 0, Endian endian = Endian.big]) =>
+      setInt64(data, n, offset, endian);
 
   @override
-  void setInt64(ByteData data, int n, [int offset = 0]) {
+  void setInt64(ByteData data, int n,
+      [int offset = 0, Endian endian = Endian.big]) {
     if (n.isNegative) {
       if (n >= -0x80000000) {
-        data.setUint32(offset, 0xFFFFFFFF);
-        data.setUint32(offset + 4, n);
+        data.setUint32(offset, 0xFFFFFFFF, endian);
+        data.setUint32(offset + 4, n, endian);
       } else {
-        _writeUint64Impl(data, n, offset);
+        _writeUint64Impl(data, n, offset, endian);
       }
     } else {
       if (n <= 0xFFFFFFFF) {
-        data.setUint32(offset, 0);
-        data.setUint32(offset + 4, n);
+        data.setUint32(offset, 0, endian);
+        data.setUint32(offset + 4, n, endian);
       } else {
-        _writeUint64Impl(data, n, offset);
+        _writeUint64Impl(data, n, offset, endian);
       }
     }
   }
 
-  void _writeUint64Impl(ByteData data, int n, int offset) {
+  void _writeUint64Impl(ByteData data, int n, int offset, Endian endian) {
     checkSafeInteger(n);
 
     // Right shift operator (>>) will cast to 32 bits in JS:
@@ -83,23 +85,37 @@ class DataSerializerPlatformGeneric extends DataSerializerPlatform {
 
     var n2 = n & 0xFFFFFFFF;
 
-    data.setUint32(offset, n1, Endian.big);
-    data.setUint32(offset + 4, n2, Endian.big);
+    if (endian == Endian.big) {
+      data.setUint32(offset, n1, Endian.big);
+      data.setUint32(offset + 4, n2, Endian.big);
+    } else {
+      data.setUint32(offset, n2, Endian.little);
+      data.setUint32(offset + 4, n1, Endian.little);
+    }
   }
 
   @override
-  int getUint64(ByteData data, [int offset = 0]) => getInt64(data, offset);
+  int getUint64(ByteData data, [int offset = 0, Endian endian = Endian.big]) =>
+      getInt64(data, offset, endian);
 
   @override
-  int getInt64(ByteData data, [int offset = 0]) {
-    var offsetN2 = offset + 4;
+  int getInt64(ByteData data, [int offset = 0, Endian endian = Endian.big]) {
+    int offsetN1, offsetN2;
 
-    var n1 = data.getUint32(offset);
+    if (endian == Endian.big) {
+      offsetN1 = offset;
+      offsetN2 = offset + 4;
+    } else {
+      offsetN1 = offset + 4;
+      offsetN2 = offset;
+    }
+
+    var n1 = data.getUint32(offsetN1, endian);
 
     if (n1 == 0) {
-      return data.getUint32(offsetN2);
+      return data.getUint32(offsetN2, endian);
     } else if (n1 == 0xFFFFFFFF) {
-      var n2 = data.getInt32(offsetN2);
+      var n2 = data.getInt32(offsetN2, endian);
       var n = (-0x800000000 - 1) + ((0x800000001) + n2);
 
       if (!n.isNegative) {
@@ -109,32 +125,33 @@ class DataSerializerPlatformGeneric extends DataSerializerPlatform {
 
       return n;
     } else if (n1 >= 0x80000000) {
-      n1 = data.getInt32(offset);
-      var n2 = data.getUint32(offsetN2);
+      var n2 = data.getUint32(offsetN2, endian);
       var n = (n1 * 0x100000000) + n2;
       return n;
     } else {
-      var n2 = data.getUint32(offsetN2);
+      var n2 = data.getUint32(offsetN2, endian);
       var n = (n1 * 0x100000000) + n2;
       return n;
     }
   }
 
   @override
-  void writeUint64(Uint8List out, int n, [int offset = 0]) =>
-      setInt64(out.asByteData(), n, offset);
+  void writeUint64(Uint8List out, int n,
+          [int offset = 0, Endian endian = Endian.big]) =>
+      setInt64(out.asByteData(), n, offset, endian);
 
   @override
-  void writeInt64(Uint8List out, int n, [int offset = 0]) =>
-      setInt64(out.asByteData(), n, offset);
+  void writeInt64(Uint8List out, int n,
+          [int offset = 0, Endian endian = Endian.big]) =>
+      setInt64(out.asByteData(), n, offset, endian);
 
   @override
-  int readUint64(Uint8List out, [int offset = 0]) =>
-      getInt64(out.asByteData(), offset);
+  int readUint64(Uint8List out, [int offset = 0, Endian endian = Endian.big]) =>
+      getInt64(out.asByteData(), offset, endian);
 
   @override
-  int readInt64(Uint8List out, [int offset = 0]) =>
-      getInt64(out.asByteData(), offset);
+  int readInt64(Uint8List out, [int offset = 0, Endian endian = Endian.big]) =>
+      getInt64(out.asByteData(), offset, endian);
 }
 
 DataSerializerPlatform createPlatformInstance() =>